requirements.txt
emoji==0.5.1
python-telegram-bot==11.1.0


